Separate Generic and Domain-Specific Logic in Context Processors
- Move ModelFittingContextProcessor to `semantiva/workflows`
