+++ semantiva/workflows/fitting_model.py
@@ -13,9 +13,13 @@
 # limitations under the License.
 
 from abc import ABC, abstractmethod
-from typing import Dict, Generic, List, TypeVar
+from typing import Dict, Generic, List, Optional, Tuple, TypeVar, Union
 
 import numpy as np
+
+from semantiva.context_processors.context_processors import ContextProcessor
+from semantiva.context_processors.context_types import ContextType
+from semantiva.logger import Logger
 
 X = TypeVar("X")  # Domain (annotation) type (e.g., float, tuple[int, float], etc.)
 Y = TypeVar("Y")  # Codomain (extracted features) type (e.g., float, list, dict, etc.)
@@ -108,6 +112,89 @@ class PolynomialFittingModel(FittingModel[float, float]):
             str: The class name and polynomial degree.
         """
         return f"{self.__class__.__name__}(degree={self.degree})"
+
+
+class ModelFittingContextProcessor(ContextProcessor):
+    """ContextProcessor that fits extracted features using a specified model."""
+
+    def __init__(
+        self,
+        logger: Optional[Logger],
+        fitting_model: FittingModel,
+        independent_var_key: str,
+        dependent_var_key: Union[str, Tuple[str, str], List[str]],
+        context_keyword: str,
+    ) -> None:
+        super().__init__(logger)
+        self.logger.debug(f"Initializing {self.__class__.__name__}")
+        self.fitting_model: FittingModel = fitting_model
+        self.independent_var_key = independent_var_key
+        self.context_keyword = context_keyword
+
+        if isinstance(dependent_var_key, (tuple, list)):
+            self.dependent_var_key = dependent_var_key[0]
+            self.dependent_var_subkey: Optional[str] = dependent_var_key[1]
+        else:
+            self.dependent_var_key = dependent_var_key
+            self.dependent_var_subkey = None
+
+        if not isinstance(independent_var_key, str):
+            raise TypeError(
+                f"independent_var_key must be a string, got {type(independent_var_key).__name__} with value {independent_var_key}"
+            )
+
+    def _process_logic(self, context: ContextType) -> ContextType:
+        """Fit extracted features to the model using context data."""
+
+        # Retrieve independent and dependent variables from context
+        independent_variable = context.get_value(self.independent_var_key)
+        dependent_variable = context.get_value(self.dependent_var_key)
+
+        # Ensure required parameters exist
+        if independent_variable is None or dependent_variable is None:
+            missing_params = [
+                p for p in self.get_required_keys() if context.get_value(p) is None
+            ]
+            raise ValueError(
+                f"Missing required context parameters: {', '.join(str(missing_params))}"
+            )
+
+        # Extract dependent_variable from dictionary if needed
+        if isinstance(self.dependent_var_subkey, tuple):
+            dependent_variable_ = tuple(
+                dependent_variable[key] for key in self.dependent_var_subkey
+            )
+        elif isinstance(self.dependent_var_subkey, str):
+            dependent_variable_ = [
+                item[self.dependent_var_subkey] for item in dependent_variable
+            ]
+        elif isinstance(dependent_variable, list):
+            dependent_variable_ = dependent_variable
+        else:
+            raise ValueError("Invalid type for dependent_variable")
+
+        # Fit the model using extracted features
+        self.logger.debug("\tRunning model %s", self.fitting_model)
+        self.logger.debug(f"\t\tindependent_variable = {independent_variable}")
+        self.logger.debug(f"\t\tdependent_variable = {dependent_variable_}")
+        fit_results = self.fitting_model.fit(independent_variable, dependent_variable_)
+
+        # Store the results back in context under the dependent variable name
+        context.set_value(self.context_keyword, fit_results)
+
+        return context
+
+    def get_required_keys(self) -> List[str]:
+        """Retrieve the list of required keys for the context operation."""
+        return [self.independent_var_key, self.dependent_var_key]
+
+    def get_created_keys(self) -> List[str]:
+        """Retrieves the list of keys created in the context."""
+        return [self.context_keyword]
+
+    def get_suppressed_keys(self) -> List[str]:
+        """This operation does not suppress any keys."""
+        return []
 
 
 # Example Usage:
